Refactor arithmetic operators into different files to reduce complexity
