+++ src/operators/arithmetic/__init__.py
@@ -1,0 +1,35 @@
+from typing import Callable, Generic, Optional, TypeVar
+from src import Statement, Type
+from src.operators import Operand
+from src.predicates import HasValidBaseTypeAndSign, IsValidArithmeticOperand
+
+
+S = TypeVar("S")
+D = TypeVar("D")
+SC = TypeVar("SC")
+DC = TypeVar("DC")
+
+
+class ArithmeticOperator(Statement, Generic[S, D, SC, DC]):
+    OPERAND_SELECTION_PREDICATE: Callable[
+        [Operand], bool
+    ] = lambda _, target_type, signed: lambda op: IsValidArithmeticOperand(
+        op
+    ) and HasValidBaseTypeAndSign(
+        op, target_type, signed
+    )
+
+
+class UnaryArithmeticOperator(
+    ArithmeticOperator[S, Optional[D], Optional[SC], Optional[DC]]
+):
+    type: Type = None
+    operand: Operand = None
+
+
+class BinaryArithmeticOperator(
+    ArithmeticOperator[S, Optional[D], Optional[SC], Optional[DC]]
+):
+    type: Type = None
+    operand1: Operand = None
+    operand2: Operand = None
