+++ src/operators/arithmetic/scalar_arithmetic.py
@@ -1,0 +1,106 @@
+from src import Signed, Unsigned
+
+from src.operators.arithmetic import BinaryArithmeticOperator, UnaryArithmeticOperator
+
+from src.operators import BinaryOperatorFuzzMixin, UnaryOperatorFuzzMixin
+
+from src.types.concrete_types import (
+    OpTypeFloat,
+    OpTypeInt,
+)
+
+
+class OpSNegate(
+    UnaryOperatorFuzzMixin, UnaryArithmeticOperator[OpTypeInt, None, Signed, None]
+):
+    ...
+
+
+class OpFNegate(
+    UnaryOperatorFuzzMixin, UnaryArithmeticOperator[OpTypeFloat, None, None, None]
+):
+    ...
+
+
+class OpIAdd(
+    BinaryOperatorFuzzMixin, BinaryArithmeticOperator[OpTypeInt, None, None, None]
+):
+    ...
+
+
+class OpFAdd(
+    BinaryOperatorFuzzMixin, BinaryArithmeticOperator[OpTypeFloat, None, None, None]
+):
+    ...
+
+
+class OpISub(
+    BinaryOperatorFuzzMixin, BinaryArithmeticOperator[OpTypeInt, None, None, None]
+):
+    ...
+
+
+class OpFSub(
+    BinaryOperatorFuzzMixin, BinaryArithmeticOperator[OpTypeFloat, None, None, None]
+):
+    ...
+
+
+class OpIMul(
+    BinaryOperatorFuzzMixin, BinaryArithmeticOperator[OpTypeInt, None, None, None]
+):
+    ...
+
+
+class OpFMul(
+    BinaryOperatorFuzzMixin, BinaryArithmeticOperator[OpTypeFloat, None, None, None]
+):
+    ...
+
+
+class OpUDiv(
+    BinaryOperatorFuzzMixin, BinaryArithmeticOperator[OpTypeInt, None, Unsigned, None]
+):
+    ...
+
+
+class OpSDiv(
+    BinaryOperatorFuzzMixin, BinaryArithmeticOperator[OpTypeInt, None, Signed, None]
+):
+    ...
+
+
+class OpFDiv(
+    BinaryOperatorFuzzMixin, BinaryArithmeticOperator[OpTypeFloat, None, None, None]
+):
+    ...
+
+
+class OpUMod(
+    BinaryOperatorFuzzMixin, BinaryArithmeticOperator[OpTypeInt, None, Unsigned, None]
+):
+    ...
+
+
+class OpSRem(
+    BinaryOperatorFuzzMixin, BinaryArithmeticOperator[OpTypeInt, None, Signed, None]
+):
+    ...
+
+
+class OpSMod(
+    BinaryOperatorFuzzMixin, BinaryArithmeticOperator[OpTypeInt, None, Signed, None]
+):
+    ...
+
+
+class OpFRem(
+    BinaryOperatorFuzzMixin, BinaryArithmeticOperator[OpTypeFloat, None, None, None]
+):
+    ...
+
+
+class OpFMod(
+    BinaryOperatorFuzzMixin, BinaryArithmeticOperator[OpTypeFloat, None, None, None]
+):
+    ...
