+++ tests/test_arithmetic.py
@@ -1,25 +1,22 @@
 import copy
-from dataclasses import replace
 import unittest
 from src import FuzzDelegator, Type
 from src.constants import (
     OpConstant,
     OpConstantComposite,
-    OpConstantFalse,
-    OpConstantTrue,
 )
 from src.monitor import Monitor
 from src.enums import ExecutionModel
 from src.context import Context
 from run_local import SPIRVSmithConfig
-from src.operators.arithmetic import (
+from src.operators.arithmetic.linear_algebra import (
     OpMatrixTimesMatrix,
     OpMatrixTimesVector,
     OpOuterProduct,
     OpVectorTimesMatrix,
     OpVectorTimesScalar,
 )
-from src.types.concrete_types import OpTypeBool, OpTypeFloat, OpTypeInt, OpTypeVector
+from src.types.concrete_types import OpTypeFloat, OpTypeVector
 
 N = 1000
 monitor = Monitor()
@@ -84,15 +81,15 @@ class TestArithmetic(unittest.TestCase):
         # Here we get either: mat2x2, mat2x4, mat4x2, mat4x4
         outer_product: OpOuterProduct = OpOuterProduct().fuzz(self.context)[-1]
         self.context.symbol_table[outer_product] = outer_product.id
-        
+
         # This should always find operands
         vector_times_matrix: OpVectorTimesMatrix = OpVectorTimesMatrix().fuzz(
             self.context
         )[-1]
-        
+
         # The resulting vector must have as many elements as the matrix has columns
         self.assertEqual(len(vector_times_matrix.type), len(outer_product.type.type))
-    
+
     def test_matrix_times_vector_has_correct_shape(self):
         create_vector_const(self.context, OpTypeFloat, size=4)
         create_vector_const(self.context, OpTypeFloat, size=2)
@@ -100,15 +97,15 @@ class TestArithmetic(unittest.TestCase):
         # Here we get either: mat2x2, mat2x4, mat4x2, mat4x4
         outer_product: OpOuterProduct = OpOuterProduct().fuzz(self.context)[-1]
         self.context.symbol_table[outer_product] = outer_product.id
-        
+
         # This should always find operands
         matrix_times_vector: OpMatrixTimesVector = OpMatrixTimesVector().fuzz(
             self.context
         )[-1]
-        
+
         # The resulting vector must have as many elements as the matrix has rows
         self.assertEqual(len(matrix_times_vector.type), len(outer_product.type.type))
-        
+
     def test_matrix_times_matrix_has_correct_shape(self):
         create_vector_const(self.context, OpTypeFloat, size=4)
         create_vector_const(self.context, OpTypeFloat, size=2)
@@ -116,20 +113,25 @@ class TestArithmetic(unittest.TestCase):
         # Here we get either: mat2x2, mat2x4, mat4x2, mat4x4
         outer_product1: OpOuterProduct = OpOuterProduct().fuzz(self.context)[-1]
         self.context.symbol_table[outer_product1] = outer_product1.id
-        
-        # We do this to avoid having two outer products that 
+
+        # We do this to avoid having two outer products that
         # eeturn a mat2x4/mat4x2 which will fail the test
         outer_product2 = copy.deepcopy(outer_product1)
         while outer_product2 == outer_product1:
             outer_product2: OpOuterProduct = OpOuterProduct().fuzz(self.context)[-1]
         self.context.symbol_table[outer_product2] = outer_product2.id
-        
+
         # This should always find operands
         matrix_times_matrix: OpMatrixTimesMatrix = OpMatrixTimesMatrix().fuzz(
             self.context
         )[-1]
-        
+
         # The resulting matrix must have as many rows as the first operand
-        self.assertEqual(len(matrix_times_matrix.type.type), len(matrix_times_matrix.operand1.type.type))
+        self.assertEqual(
+            len(matrix_times_matrix.type.type),
+            len(matrix_times_matrix.operand1.type.type),
+        )
         # The resulting matrix must have as many columns as the second operand
-        self.assertEqual(len(matrix_times_matrix.type.type), len(matrix_times_matrix.operand2.type))
+        self.assertEqual(
+            len(matrix_times_matrix.type.type), len(matrix_times_matrix.operand2.type)
+        )
